Fix encoding issue
If the lines of a log have different encodings and those encodings are not
automatically joinable by Ruby then Ruby will raise an
`Encoding::CompatibilityError`.  See the added spec.

StringIO seems to have the code require to merge the various encodings so this
uses a StringIO as the buffer rather than an array.